Pre-empting merge conflicts with pr 21 create a functional api
